homicides/requirements.txt
Django==1.10.4
django-filter==1.0.1
